Second ar page created

+++ cartissimo/lib/pages/player_detail.dart
@@ -1,13 +1,11 @@
+import 'package:cartissimo/models/race.dart';
 import 'package:cartissimo/pages/ar_page_customize_cart.dart';
+import 'package:cartissimo/pages/ar_page_race.dart';
 import 'package:flutter/material.dart';
 import '../models/player.dart';
 import '../apis/player_api.dart';
+import '../apis/race_api.dart';
 
-const List<String> choices = <String>[
-  'Save Player & Back',
-  'Delete Player',
-  'Back to List'
-];
 
 class PlayerDetailPage extends StatefulWidget {
   final int id; // our UserDetailPage has an id-parameter which contains the id of the user to show
@@ -22,6 +20,7 @@ class PlayerDetailPage extends StatefulWidget {
 
 class _PlayerDetailPageState extends State<PlayerDetailPage> {
   Player? player;
+  List<Race> races = [];
 
   TextEditingController nameController = TextEditingController();
 
@@ -32,6 +31,7 @@ class _PlayerDetailPageState extends State<PlayerDetailPage> {
       player = Player(id: 0, name: "", optionIndex: 0);
     } else {
       _getPlayer(widget.id);
+      _getRaces();
     }
   }
 
@@ -42,6 +42,10 @@ class _PlayerDetailPageState extends State<PlayerDetailPage> {
         player = result;
       });
     });
+  }
+
+  void _getRaces() {
+    RaceApi.fetchRaces().then((value) => races = value);
   }
 
   @override
@@ -63,100 +67,121 @@ class _PlayerDetailPageState extends State<PlayerDetailPage> {
 
     nameController.text = player!.name; // Set the initial name in the TextField
 
-    return Padding(
-  padding: const EdgeInsets.all(10.0),
-  child: Column(
-    children: <Widget>[
-      // Add a button to go back at the top left
-      Align(
-        alignment: Alignment.topLeft,
-        child: ElevatedButton.icon(
-          onPressed: () {
-            Navigator.pop(context); // Go back to the previous page
-          },
-          icon: const Icon(Icons.arrow_back),
-          label: const Text("Back"),
-          style: ElevatedButton.styleFrom(
-            backgroundColor: Colors.grey, // Optional: change button color
-          ),
-        ),
-      ),
-      const SizedBox(height: 20), // Add spacing between the top button and the form
-      TextField(
-        controller: nameController,
-        style: textStyle,
-        keyboardType: TextInputType.text,
-        decoration: InputDecoration(
-          labelText: "Name",
-          labelStyle: textStyle,
-          border: OutlineInputBorder(
-            borderRadius: BorderRadius.circular(5.0),
-          ),
-        ),
-      ),
-      const SizedBox(height: 20), // Add space between the TextField and buttons
-      Row(
-        mainAxisAlignment: MainAxisAlignment.start,
+    return Padding(    
+      padding: const EdgeInsets.all(10.0),
+      child: Column(
         children: <Widget>[
-          Container(
-            margin: const EdgeInsets.all(10),
-            child: ElevatedButton(
-              onPressed: _savePlayer,
-              child: Text(player!.id == 0 ? "Add" : "Update"),
+          // Add a button to go back at the top left
+          Align(
+            alignment: Alignment.topLeft,
+            child: ElevatedButton.icon(
+              onPressed: () {
+                Navigator.pop(context); // Go back to the previous page
+              },
+              icon: const Icon(Icons.arrow_back),
+              label: const Text("Back"),
+              style: ElevatedButton.styleFrom(
+                backgroundColor: Colors.grey, // Optional: change button color
+              ),
             ),
           ),
-          Container(
-            margin: const EdgeInsets.all(10),
-            child: ElevatedButton(
-              onPressed: () {
-                _confirmDelete();
-              },
-              style: ElevatedButton.styleFrom(backgroundColor: Colors.red),
-              child: const Text("Delete"),
+          const SizedBox(height: 20), // Add spacing between the top button and the form
+          TextField(
+            controller: nameController,
+            style: textStyle,
+            keyboardType: TextInputType.text,
+            decoration: InputDecoration(
+              labelText: "Name",
+              labelStyle: textStyle,
+              border: OutlineInputBorder(
+                borderRadius: BorderRadius.circular(5.0),
+              ),
             ),
-          )
-        ],
-      ),
-      const SizedBox(height: 20), 
-      Row(
-        mainAxisAlignment: MainAxisAlignment.start,
-        children: [
-          Container(
-            margin: const EdgeInsets.all(10),
-            child: ElevatedButton(
-              onPressed: () {
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(builder: (context) => ArPageCustomizeCart(player: player!)), // Replace `DifferentPage` with the actual page widget
-                );
-              },
-              child: const Text("Customize Cart"),
-            ),
-          )
-        ],
-      ),
-      
-    ],
-  ),
-);
+          ),
+          const SizedBox(height: 20), // Add space between the TextField and buttons
+          Row(
+            mainAxisAlignment: MainAxisAlignment.start,
+            children: <Widget>[
+              Container(
+                margin: const EdgeInsets.all(10),
+                child: ElevatedButton(
+                  onPressed: _savePlayer,
+                  child: Text(player!.id == 0 ? "Add" : "Update"),
+                ),
+              ),
+              Container(
+                margin: const EdgeInsets.all(10),
+                child: ElevatedButton(
+                  onPressed: () {
+                    _confirmDelete(player!.id);
+                  },
+                  style: ElevatedButton.styleFrom(backgroundColor: Colors.red),
+                  child: Text(player!.id == 0? "Cancel": "Delete"),
+                ),
+              )
+            ],
+          ),
+          const SizedBox(height: 20), 
+          Column(
+            crossAxisAlignment: CrossAxisAlignment.start,  // Align children to the left
+            children: [
+              // The button will now be aligned to the left
+              Container(
+                margin: const EdgeInsets.all(10),
+                child: ElevatedButton(
+                  onPressed: () {
+                    Navigator.push(
+                      context,
+                      MaterialPageRoute(builder: (context) => ArPageCustomizeCart(player: player!)),
+                    );
+                  },
+                  child: const Text("Customize Cart"),
+                ),
+              ),
+              
+              const SizedBox(height: 20), // Add space between button and ListView
 
+              // The ListView will be below the button
+              Container(
+                margin: const EdgeInsets.all(10),
+                child: ListView.builder(
+                  shrinkWrap: true, // Ensures ListView only takes as much space as needed
+                  physics: const NeverScrollableScrollPhysics(), // Disables independent scrolling of the ListView
+                  itemCount: races.length,
+                  itemBuilder: (context, index) {
+                    final race = races[index];
+                    return ListTile(
+                      title: Text(race.name),
+                      subtitle: Text(
+                        "Difficulty: ${race.difficulty.name}, Distance: ${race.distance}m, Time Limit: ${race.beforeTimeLimit} minutes",
+                      ),
+                      trailing: ElevatedButton(
+                        onPressed: () {
+                          // Perform action when race is selected
+                          _selectRace(context, race);
+                        },
+                        child: const Text("Select"),
+                      ),
+                    );
+                  },
+                ),
+              ),
+            ],
+          ),
+
+
+
+        ],
+      ),
+    );
   }
-}
+  }
 
-
-  void _menuSelected(String index) async {
-    switch (index) {
-      case "0": // Save User & Back
-        _savePlayer();
-        break;
-      case "1": // Delete User
-        _deletePlayer();
-        break;
-      case "2": // Back to List
-        Navigator.pop(context, true);
-        break;
-      default:
-    }
+  void _selectRace(BuildContext context, Race race) {
+    Navigator.push(
+      context,
+      MaterialPageRoute(builder: (context) => ArPageRace(race: race)),
+    );
   }
 
   void _savePlayer() {
@@ -179,30 +204,35 @@ class _PlayerDetailPageState extends State<PlayerDetailPage> {
     });
   }
 
-  void _confirmDelete() {
-  showDialog(
-    context: context,
-    builder: (BuildContext context) {
-      return AlertDialog(
-        title: const Text('Confirm Deletion'),
-        content: const Text('Are you sure you want to delete this player?'),
-        actions: [
-          TextButton(
-            child: const Text('Cancel'),
-            onPressed: () => Navigator.of(context).pop(),
-          ),
-          TextButton(
-            child: const Text('Delete'),
-            onPressed: () {
-              Navigator.of(context).pop(); // Close the dialog
-              _deletePlayer(); // Perform the deletion
-            },
-          ),
-        ],
-      );
-    },
-  );
-}
+  void _confirmDelete(int id) {
+    if (id == 0) {
+      Navigator.pop(context, true);
+    } else {
+      showDialog(
+      context: context,
+      builder: (BuildContext context) {
+        return AlertDialog(
+          title: const Text('Confirm Deletion'),
+          content: const Text('Are you sure you want to delete this player?'),
+          actions: [
+            TextButton(
+              child: const Text('Cancel'),
+              onPressed: () => Navigator.of(context).pop(),
+            ),
+            TextButton(
+              child: const Text('Delete'),
+              onPressed: () {
+                Navigator.of(context).pop(); // Close the dialog
+                _deletePlayer(); // Perform the deletion
+              },
+            ),
+          ],
+        );
+      },
+    );
+    }
+    
+  }
 
 
 }
